Test Registry and EventStore bus_driver implementations
